caca/requirements.txt
numpy
opencv
pyzmq
